+++ NEWS.md
@@ -1,3 +1,7 @@
+# Changes in version 0.2.1 (2024-09-27)
+* Support for plotly zoom and scroll
+* Support for using saved image backgrounds
+
 # Changes in version 0.2.0 (2024-09-20)
 * Support for interactive plot conversion
 * Warning when original colors are too similar